using cast size_t to int
Former-commit-id: 6eebebecfad8cf2957bcbdf0d5075e6123f7224d

+++ src/AudioModel/AudioModel.h
@@ -28,7 +28,7 @@
 
 class AudioModel {
 public:
-    AudioModel(double *x, int x_length, int fs, const lessConfigure &configure);
+    AudioModel(double *x, size_t x_length, int fs, const lessConfigure &configure);
 
     lessAudioModel GetAudioModel();
